various improvements

+++ web/Laravel/laravel/shell.nix
@@ -10,14 +10,28 @@ in
     nativeBuildInputs = [ appPackage.nodeShell ];
     inputsFrom = [ appPackage ];
     src = null;
-
-    # FIXME for laravel we must copy vendor dir and run composer dump-autoload
     shellHook = ''
-      [ ! -e vendor ] && mkdir vendor && cp -R ${appPackage.phpDeps}/* vendor
-      chmod -R u+w vendor
-      composer dump-autoload
-      ln -s ${appPackage.nodePackage}/lib/node_modules/laravel-node-dependencies/node_modules node_modules
+      if [ ! -e vendor ] ; then
+        mkdir -p vendor/composer
+        cp ${appPackage.phpDeps}/composer/* vendor/composer/
+        shopt -s extglob
+        ln -s ${appPackage.phpDeps}/!(composer) vendor/
+        composer dump-autoload
+        PHP_DEPENDENCIES_LINKED=true
+      fi
+
+      if [ ! -e node_modules ] ; then
+        ln -s ${appPackage.nodePackage}/lib/node_modules/laravel-node-dependencies/node_modules node_modules
+        NODE_DEPENDENCIES_LINKED=true
+      fi
+
       alias cross-env='true &&'
-      trap "rm -rf vendor node_modules" EXIT
+
+      exitHandler () {
+          [ ! -z $PHP_DEPENDENCIES_LINKED ] && rm -rf vendor
+          [ ! -z $NODE_DEPENDENCIES_LINKED ] && rm node_modules
+      }
+
+      trap exitHandler EXIT
     '';
   }
